Update image metadata, add manual sync endpoint

+++ lib/Lightroom/Responses/Exif.cs
@@ -2,8 +2,8 @@ namespace ldam.co.za.lib.Lightroom
 {
     public class Exif
     {
-        public decimal[] ExposureTime { get; set; }
-        public decimal[] FNumber { get; set; }
+        public int[] ExposureTime { get; set; }
+        public int[] FNumber { get; set; }
         public int[] FocalLength { get; set; }
         public int ISOSpeedRatings { get; set; }
     }
